Restructure subspace-networking crate into several modules
