+++ crates/subspace-networking/Cargo.toml
@@ -14,6 +14,7 @@ include = [
 
 [dependencies]
 event-listener-primitives = "2.0.1"
+futures = "0.3.19"
 log = "0.4.14"
 parking_lot = "0.12.0"
 thiserror = "1.0.30"
